business/chongqing_zhaoshang/source/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
lxml==5.2.1
multidict==6.0.5
pyBrainai==1.4.0.9
python-docx==1.1.0
requests==2.31.0
typing_extensions==4.10.0
urllib3==2.2.1
yarl==1.9.4